resource-timing/event-source-timing.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing: EventSource timing behavior</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resource-timing/resources/observe-entry.js"></script>
</head>
</script>
<script>
    async_test(t => {
        const delay = 100;
        const repetitions = 4;
        const url = new URL(`/eventsource/resources/message.py?sleep=${delay}`, location.href);
        const eventSource = new EventSource(url);
        let messages = 0;
        let entries = 0;
        eventSource.addEventListener('message', () => {
            ++messages;
        })

        new PerformanceObserver(() => {
            entries = performance.getEntriesByName(url).length;
            if (entries === repetitions) {
                assert_equals(messages, repetitions);
                t.done();
            }
        }).observe({type: 'resource'});
    }, "ResourceTiming for EventSource should reflect number of re-connections to source");
</script>
</body>
</html>
